# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuditMode = (typeof AuditMode)[keyof typeof AuditMode];

export const AuditMode = {
  code: "code",
  dependency: "dependency",
} as const;

export interface SecurityAuditRequest {
  /**
   * @minLength 1
   * @maxLength 20000
   */
  code: string;
  /**
   * Optional language hint such as JavaScript, Python, SQL, Go, or Bash.
   * @maxLength 80
   */
  language?: string;
  mode?: AuditMode;
}

export type Severity = (typeof Severity)[keyof typeof Severity];

export const Severity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Vulnerability {
  type: string;
  severity: Severity;
  /** @minimum 1 */
  line: number;
  location: string;
  evidence: string;
  remediation: string;
}

export interface SecureRewrite {
  vulnerable: string;
  hardened: string;
  notes: string;
}

export type DependencyStatus =
  (typeof DependencyStatus)[keyof typeof DependencyStatus];

export const DependencyStatus = {
  safe: "safe",
  outdated: "outdated",
  vulnerable: "vulnerable",
  unknown: "unknown",
} as const;

export interface DependencyCve {
  id: string;
  severity: Severity;
  description: string;
}

export interface DependencyFinding {
  name: string;
  currentVersion: string;
  ecosystem: string;
  status: DependencyStatus;
  safeVersion?: string;
  advisory?: string;
  cves: DependencyCve[];
}

export interface AttackVector {
  narrative: string;
  attackerProfile: string;
  steps: string[];
  proofOfConcept: string;
  pocLanguage: string;
  impact: string;
  mitigation: string;
}

export interface SecurityAuditResult {
  summary: string;
  severity: Severity;
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
  hardened: boolean;
  badge: string;
  mode: AuditMode;
  vulnerabilities: Vulnerability[];
  secureRewrite: SecureRewrite;
  checklist: string[];
  attackVector: AttackVector;
  dependencies?: DependencyFinding[];
}

export type AuditChatMessageRole =
  (typeof AuditChatMessageRole)[keyof typeof AuditChatMessageRole];

export const AuditChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AuditChatMessage {
  role: AuditChatMessageRole;
  /**
   * @minLength 1
   * @maxLength 4000
   */
  content: string;
}

export interface AuditChatContext {
  /** @maxLength 20000 */
  code: string;
  /** @maxLength 80 */
  language?: string;
  mode?: AuditMode;
  /** @maxLength 4000 */
  summary?: string;
  vulnerabilities?: Vulnerability[];
  dependencies?: DependencyFinding[];
}

export interface AuditChatRequest {
  /**
   * @minItems 1
   * @maxItems 30
   */
  messages: AuditChatMessage[];
  context: AuditChatContext;
}

export interface AuditChatResponse {
  reply: string;
}

export interface AuditRulesResponse {
  rules: string[];
}

export interface ErrorResponse {
  message: string;
}

export interface RateLimitErrorResponse {
  message: string;
  /**
   * Seconds to wait before retrying.
   * @minimum 0
   */
  retryAfter: number;
}

/**
 * Headless audit payload. Provide either 'code' or 'file' (synonyms).
 */
export interface CliAuditRequest {
  /**
   * @minLength 1
   * @maxLength 20000
   */
  code?: string;
  /**
   * @minLength 1
   * @maxLength 20000
   */
  file?: string;
  /** @maxLength 200 */
  filename?: string;
  /** @maxLength 80 */
  language?: string;
}

export type CliCriticalVulnerabilitySeverity =
  (typeof CliCriticalVulnerabilitySeverity)[keyof typeof CliCriticalVulnerabilitySeverity];

export const CliCriticalVulnerabilitySeverity = {
  high: "high",
  critical: "critical",
} as const;

export interface CliCriticalVulnerability {
  type: string;
  severity: CliCriticalVulnerabilitySeverity;
  /** @minimum 1 */
  line: number;
  evidence: string;
  remediation: string;
}

export interface CliAuditResponse {
  /**
   * @minimum 0
   * @maximum 100
   */
  security_score: number;
  critical_vulnerabilities: CliCriticalVulnerability[];
}

export interface ProjectTokenSummary {
  /** @minimum 1 */
  id: number;
  name: string;
  prefix: string;
  createdAt: string;
  lastUsedAt: string | null;
  /** @minimum 0 */
  requestCount: number;
  revoked: boolean;
}

export interface ListProjectTokensResponse {
  tokens: ProjectTokenSummary[];
}

export interface CreateProjectTokenRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
}

export interface CreateProjectTokenResponse {
  /** @minimum 1 */
  id: number;
  name: string;
  prefix: string;
  createdAt: string;
  token: string;
  notice: string;
}

export interface RevokeProjectTokenResponse {
  /** @minimum 1 */
  id: number;
  revoked: boolean;
}