# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: audit
    description: Sentinel security audit operations
  - name: cli
    description: Headless CLI / CI integration endpoints
  - name: tokens
    description: Project token administration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /audits:
    post:
      operationId: createSecurityAudit
      tags: [audit]
      summary: Analyze a code snippet for security flaws
      description: Runs a security-first logic audit and returns specific vulnerabilities, remediation guidance, and a hardened rewrite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityAuditRequest"
      responses:
        "200":
          description: Security audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityAuditResult"
        "400":
          description: Invalid audit request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Audit failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /audits/chat:
    post:
      operationId: chatWithAudit
      tags: [audit]
      summary: Ask follow-up questions about a Sentinel audit
      description: Continues a conversation grounded in the most recent Sentinel audit context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditChatRequest"
      responses:
        "200":
          description: Sentinel chat reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditChatResponse"
        "400":
          description: Invalid chat request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Chat failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /audit-rules:
    get:
      operationId: getAuditRules
      tags: [audit]
      summary: Get Sentinel audit focus areas
      description: Returns the vulnerability classes Sentinel emphasizes during analysis.
      responses:
        "200":
          description: Audit focus areas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRulesResponse"
  /audit-cli:
    post:
      operationId: runCliAudit
      tags: [cli]
      summary: Headless CLI / CI security audit
      description: Token-gated audit endpoint for git hooks, pre-commit, and CI pipelines. Returns only the strict security score and high or critical vulnerabilities.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliAuditRequest"
      responses:
        "200":
          description: CLI audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CliAuditResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid project token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload exceeds maximum allowed size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded for this token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitErrorResponse"
        "500":
          description: Audit failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tokens:
    get:
      operationId: listProjectTokens
      tags: [tokens]
      summary: List provisioned project tokens
      description: Admin-only listing of provisioned project tokens (raw token values are never returned).
      security:
        - adminAuth: []
      responses:
        "200":
          description: Provisioned tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProjectTokensResponse"
        "401":
          description: Missing or invalid admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin auth not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProjectToken
      tags: [tokens]
      summary: Provision a new project token
      description: Provisions a new sntl_ token. The raw value is returned exactly once and only its SHA-256 hash is persisted.
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectTokenRequest"
      responses:
        "201":
          description: Token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProjectTokenResponse"
        "400":
          description: Invalid project name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin auth not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tokens/{id}:
    delete:
      operationId: revokeProjectToken
      tags: [tokens]
      summary: Revoke a provisioned project token
      description: Marks the token as revoked. Subsequent CLI requests using it will return 401.
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeProjectTokenResponse"
        "401":
          description: Missing or invalid admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin auth not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sntl_token
      description: Per-project Sentinel token sent as 'Authorization Bearer sntl_...'
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Per-project Sentinel token sent as 'X-API-Key sntl_...'
    adminAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Admin token from the ADMIN_TOKEN environment variable, required to manage project tokens.
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuditMode:
      type: string
      enum: [code, dependency]
    SecurityAuditRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 20000
        language:
          type: string
          description: Optional language hint such as JavaScript, Python, SQL, Go, or Bash.
          maxLength: 80
        mode:
          $ref: "#/components/schemas/AuditMode"
      required:
        - code
    Severity:
      type: string
      enum: [low, medium, high, critical]
    Vulnerability:
      type: object
      properties:
        type:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        line:
          type: integer
          minimum: 1
        location:
          type: string
        evidence:
          type: string
        remediation:
          type: string
      required:
        - type
        - severity
        - line
        - location
        - evidence
        - remediation
    SecureRewrite:
      type: object
      properties:
        vulnerable:
          type: string
        hardened:
          type: string
        notes:
          type: string
      required:
        - vulnerable
        - hardened
        - notes
    DependencyStatus:
      type: string
      enum: [safe, outdated, vulnerable, unknown]
    DependencyCve:
      type: object
      properties:
        id:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        description:
          type: string
      required:
        - id
        - severity
        - description
    DependencyFinding:
      type: object
      properties:
        name:
          type: string
        currentVersion:
          type: string
        ecosystem:
          type: string
        status:
          $ref: "#/components/schemas/DependencyStatus"
        safeVersion:
          type: string
        advisory:
          type: string
        cves:
          type: array
          items:
            $ref: "#/components/schemas/DependencyCve"
      required:
        - name
        - currentVersion
        - ecosystem
        - status
        - cves
    AttackVector:
      type: object
      properties:
        narrative:
          type: string
        attackerProfile:
          type: string
        steps:
          type: array
          items:
            type: string
        proofOfConcept:
          type: string
        pocLanguage:
          type: string
        impact:
          type: string
        mitigation:
          type: string
      required:
        - narrative
        - attackerProfile
        - steps
        - proofOfConcept
        - pocLanguage
        - impact
        - mitigation
    SecurityAuditResult:
      type: object
      properties:
        summary:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        score:
          type: integer
          minimum: 0
          maximum: 100
        hardened:
          type: boolean
        badge:
          type: string
        mode:
          $ref: "#/components/schemas/AuditMode"
        vulnerabilities:
          type: array
          items:
            $ref: "#/components/schemas/Vulnerability"
        secureRewrite:
          $ref: "#/components/schemas/SecureRewrite"
        checklist:
          type: array
          items:
            type: string
        attackVector:
          $ref: "#/components/schemas/AttackVector"
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/DependencyFinding"
      required:
        - summary
        - severity
        - score
        - hardened
        - badge
        - mode
        - vulnerabilities
        - secureRewrite
        - checklist
        - attackVector
    AuditChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
          maxLength: 4000
      required:
        - role
        - content
    AuditChatContext:
      type: object
      properties:
        code:
          type: string
          maxLength: 20000
        language:
          type: string
          maxLength: 80
        mode:
          $ref: "#/components/schemas/AuditMode"
        summary:
          type: string
          maxLength: 4000
        vulnerabilities:
          type: array
          items:
            $ref: "#/components/schemas/Vulnerability"
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/DependencyFinding"
      required:
        - code
    AuditChatRequest:
      type: object
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/AuditChatMessage"
        context:
          $ref: "#/components/schemas/AuditChatContext"
      required:
        - messages
        - context
    AuditChatResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
    AuditRulesResponse:
      type: object
      properties:
        rules:
          type: array
          items:
            type: string
      required:
        - rules
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RateLimitErrorResponse:
      type: object
      properties:
        message:
          type: string
        retryAfter:
          type: integer
          minimum: 0
          description: Seconds to wait before retrying.
      required:
        - message
        - retryAfter
    CliAuditRequest:
      type: object
      description: Headless audit payload. Provide either 'code' or 'file' (synonyms).
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 20000
        file:
          type: string
          minLength: 1
          maxLength: 20000
        filename:
          type: string
          maxLength: 200
        language:
          type: string
          maxLength: 80
    CliCriticalVulnerability:
      type: object
      properties:
        type:
          type: string
        severity:
          type: string
          enum: [high, critical]
        line:
          type: integer
          minimum: 1
        evidence:
          type: string
        remediation:
          type: string
      required:
        - type
        - severity
        - line
        - evidence
        - remediation
    CliAuditResponse:
      type: object
      properties:
        security_score:
          type: integer
          minimum: 0
          maximum: 100
        critical_vulnerabilities:
          type: array
          items:
            $ref: "#/components/schemas/CliCriticalVulnerability"
      required:
        - security_score
        - critical_vulnerabilities
    ProjectTokenSummary:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        prefix:
          type: string
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        requestCount:
          type: integer
          minimum: 0
        revoked:
          type: boolean
      required:
        - id
        - name
        - prefix
        - createdAt
        - lastUsedAt
        - requestCount
        - revoked
    ListProjectTokensResponse:
      type: object
      properties:
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/ProjectTokenSummary"
      required:
        - tokens
    CreateProjectTokenRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
      required:
        - name
    CreateProjectTokenResponse:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        prefix:
          type: string
        createdAt:
          type: string
          format: date-time
        token:
          type: string
        notice:
          type: string
      required:
        - id
        - name
        - prefix
        - createdAt
        - token
        - notice
    RevokeProjectTokenResponse:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        revoked:
          type: boolean
      required:
        - id
        - revoked