# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Runs a security-first logic audit and returns specific vulnerabilities, remediation guidance, and a hardened rewrite.
 * @summary Analyze a code snippet for security flaws
 */
export const createSecurityAuditBodyCodeMax = 20000;

export const createSecurityAuditBodyLanguageMax = 80;

export const CreateSecurityAuditBody = zod.object({
  code: zod.string().min(1).max(createSecurityAuditBodyCodeMax),
  language: zod
    .string()
    .max(createSecurityAuditBodyLanguageMax)
    .optional()
    .describe(
      "Optional language hint such as JavaScript, Python, SQL, Go, or Bash.",
    ),
  mode: zod.enum(["code", "dependency"]).optional(),
});

export const createSecurityAuditResponseScoreMin = 0;
export const createSecurityAuditResponseScoreMax = 100;

export const CreateSecurityAuditResponse = zod.object({
  summary: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  score: zod
    .number()
    .min(createSecurityAuditResponseScoreMin)
    .max(createSecurityAuditResponseScoreMax),
  hardened: zod.boolean(),
  badge: zod.string(),
  mode: zod.enum(["code", "dependency"]),
  vulnerabilities: zod.array(
    zod.object({
      type: zod.string(),
      severity: zod.enum(["low", "medium", "high", "critical"]),
      line: zod.number().min(1),
      location: zod.string(),
      evidence: zod.string(),
      remediation: zod.string(),
    }),
  ),
  secureRewrite: zod.object({
    vulnerable: zod.string(),
    hardened: zod.string(),
    notes: zod.string(),
  }),
  checklist: zod.array(zod.string()),
  attackVector: zod.object({
    narrative: zod.string(),
    attackerProfile: zod.string(),
    steps: zod.array(zod.string()),
    proofOfConcept: zod.string(),
    pocLanguage: zod.string(),
    impact: zod.string(),
    mitigation: zod.string(),
  }),
  dependencies: zod
    .array(
      zod.object({
        name: zod.string(),
        currentVersion: zod.string(),
        ecosystem: zod.string(),
        status: zod.enum(["safe", "outdated", "vulnerable", "unknown"]),
        safeVersion: zod.string().optional(),
        advisory: zod.string().optional(),
        cves: zod.array(
          zod.object({
            id: zod.string(),
            severity: zod.enum(["low", "medium", "high", "critical"]),
            description: zod.string(),
          }),
        ),
      }),
    )
    .optional(),
});

/**
 * Continues a conversation grounded in the most recent Sentinel audit context.
 * @summary Ask follow-up questions about a Sentinel audit
 */
export const chatWithAuditBodyMessagesItemContentMax = 4000;

export const chatWithAuditBodyMessagesMax = 30;

export const chatWithAuditBodyContextCodeMax = 20000;

export const chatWithAuditBodyContextLanguageMax = 80;

export const chatWithAuditBodyContextSummaryMax = 4000;

export const ChatWithAuditBody = zod.object({
  messages: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod
          .string()
          .min(1)
          .max(chatWithAuditBodyMessagesItemContentMax),
      }),
    )
    .min(1)
    .max(chatWithAuditBodyMessagesMax),
  context: zod.object({
    code: zod.string().max(chatWithAuditBodyContextCodeMax),
    language: zod.string().max(chatWithAuditBodyContextLanguageMax).optional(),
    mode: zod.enum(["code", "dependency"]).optional(),
    summary: zod.string().max(chatWithAuditBodyContextSummaryMax).optional(),
    vulnerabilities: zod
      .array(
        zod.object({
          type: zod.string(),
          severity: zod.enum(["low", "medium", "high", "critical"]),
          line: zod.number().min(1),
          location: zod.string(),
          evidence: zod.string(),
          remediation: zod.string(),
        }),
      )
      .optional(),
    dependencies: zod
      .array(
        zod.object({
          name: zod.string(),
          currentVersion: zod.string(),
          ecosystem: zod.string(),
          status: zod.enum(["safe", "outdated", "vulnerable", "unknown"]),
          safeVersion: zod.string().optional(),
          advisory: zod.string().optional(),
          cves: zod.array(
            zod.object({
              id: zod.string(),
              severity: zod.enum(["low", "medium", "high", "critical"]),
              description: zod.string(),
            }),
          ),
        }),
      )
      .optional(),
  }),
});

export const ChatWithAuditResponse = zod.object({
  reply: zod.string(),
});

/**
 * Returns the vulnerability classes Sentinel emphasizes during analysis.
 * @summary Get Sentinel audit focus areas
 */
export const GetAuditRulesResponse = zod.object({
  rules: zod.array(zod.string()),
});

/**
 * Token-gated audit endpoint for git hooks, pre-commit, and CI pipelines. Returns only the strict security score and high or critical vulnerabilities.
 * @summary Headless CLI / CI security audit
 */
export const runCliAuditBodyCodeMax = 20000;

export const runCliAuditBodyFileMax = 20000;

export const runCliAuditBodyFilenameMax = 200;

export const runCliAuditBodyLanguageMax = 80;

export const RunCliAuditBody = zod
  .object({
    code: zod.string().min(1).max(runCliAuditBodyCodeMax).optional(),
    file: zod.string().min(1).max(runCliAuditBodyFileMax).optional(),
    filename: zod.string().max(runCliAuditBodyFilenameMax).optional(),
    language: zod.string().max(runCliAuditBodyLanguageMax).optional(),
  })
  .describe(
    "Headless audit payload. Provide either 'code' or 'file' (synonyms).",
  );

export const runCliAuditResponseSecurityScoreMin = 0;
export const runCliAuditResponseSecurityScoreMax = 100;

export const RunCliAuditResponse = zod.object({
  security_score: zod
    .number()
    .min(runCliAuditResponseSecurityScoreMin)
    .max(runCliAuditResponseSecurityScoreMax),
  critical_vulnerabilities: zod.array(
    zod.object({
      type: zod.string(),
      severity: zod.enum(["high", "critical"]),
      line: zod.number().min(1),
      evidence: zod.string(),
      remediation: zod.string(),
    }),
  ),
});

/**
 * Admin-only listing of provisioned project tokens (raw token values are never returned).
 * @summary List provisioned project tokens
 */

export const listProjectTokensResponseTokensItemRequestCountMin = 0;

export const ListProjectTokensResponse = zod.object({
  tokens: zod.array(
    zod.object({
      id: zod.number().min(1),
      name: zod.string(),
      prefix: zod.string(),
      createdAt: zod.coerce.date(),
      lastUsedAt: zod.coerce.date().nullable(),
      requestCount: zod
        .number()
        .min(listProjectTokensResponseTokensItemRequestCountMin),
      revoked: zod.boolean(),
    }),
  ),
});

/**
 * Provisions a new sntl_ token. The raw value is returned exactly once and only its SHA-256 hash is persisted.
 * @summary Provision a new project token
 */
export const createProjectTokenBodyNameMax = 100;

export const CreateProjectTokenBody = zod.object({
  name: zod.string().min(1).max(createProjectTokenBodyNameMax),
});

/**
 * Marks the token as revoked. Subsequent CLI requests using it will return 401.
 * @summary Revoke a provisioned project token
 */

export const RevokeProjectTokenParams = zod.object({
  id: zod.coerce.number().min(1),
});

export const RevokeProjectTokenResponse = zod.object({
  id: zod.number().min(1),
  revoked: zod.boolean(),
});