# lib/api-zod/src/generated/types/cliAuditRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Headless audit payload. Provide either 'code' or 'file' (synonyms).
 */
export interface CliAuditRequest {
  /**
   * @minLength 1
   * @maxLength 20000
   */
  code?: string;
  /**
   * @minLength 1
   * @maxLength 20000
   */
  file?: string;
  /** @maxLength 200 */
  filename?: string;
  /** @maxLength 80 */
  language?: string;
}