# lib/api-zod/src/generated/types/cliAuditResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CliCriticalVulnerability } from "./cliCriticalVulnerability";

export interface CliAuditResponse {
  /**
   * @minimum 0
   * @maximum 100
   */
  security_score: number;
  critical_vulnerabilities: CliCriticalVulnerability[];
}