# lib/api-zod/src/generated/types/createProjectTokenRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectTokenRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
}