# lib/api-zod/src/generated/types/createProjectTokenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectTokenResponse {
  /** @minimum 1 */
  id: number;
  name: string;
  prefix: string;
  createdAt: Date;
  token: string;
  notice: string;
}