# lib/api-zod/src/generated/types/projectTokenSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectTokenSummary {
  /** @minimum 1 */
  id: number;
  name: string;
  prefix: string;
  createdAt: Date;
  lastUsedAt: Date | null;
  /** @minimum 0 */
  requestCount: number;
  revoked: boolean;
}