# lib/api-zod/src/generated/types/rateLimitErrorResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RateLimitErrorResponse {
  message: string;
  /**
   * Seconds to wait before retrying.
   * @minimum 0
   */
  retryAfter: number;
}