# lib/api-zod/src/generated/types/revokeProjectTokenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RevokeProjectTokenResponse {
  /** @minimum 1 */
  id: number;
  revoked: boolean;
}